+++ script/patches/decaffeinated-bundled-packages/autocomplete-chevron-api/lib/main.js
@@ -1,7 +1,0 @@
-const provider = require('./provider');
-
-module.exports = {
-  activate() { return provider.load(); },
-
-  getProvider() { return provider; }
-};